put div intrinsics in their own module + some docs
